requirements.txt
dataclasses-json
flask_cors
flask-jwt-extended
flask-marshmallow
pdfkit
